Removed the hash code.
It just became painfully obvious that I didn't need to use a hash at
all.  In doing so, I create a SortingItem, and this item will house the
StringLinkedList (which will be changed shortly) and the original index
of the item.  So once that's sorted, I use that data to resort the list