Change StringValuePtr to StringValueCStr
StringValuePtr doesn't guarantee a NULL byte at the end of the char * it
returns.  The for loop in the `parse_version_number` depends on a NULL
byte in the string in order to stop the loop.  Since `StringValuePtr`
doesn't guarantee a NULL byte in the `char *`, it's possible the for
loop could read past the end of the string, and `offset` would end up
being larger than the number of bytes that are actually in the string.